feat: bundles api, schema, docs and refactor
