+++ api/test/bundles/fixtures/basic_kbda2-1.1.keyboard_info
@@ -1,0 +1,47 @@
+{
+  "license": "mit",
+  "description": "Arabic (102) Basic generated from template",
+  "related": {
+    "arabic_102": {
+      "deprecates": true
+    },
+    "kbda2": {
+      "deprecates": true
+    }
+  },
+  "id": "basic_kbda2",
+  "languages": {
+    "ar": {
+      "displayName": "Arabic",
+      "languageName": "Arabic"
+    }
+  },
+  "name": "Arabic (102) Basic",
+  "isRTL": true,
+  "lastModifiedDate": "2019-09-01T22:19:09.055Z",
+  "sourcePath": "release/basic/basic_kbda2",
+  "version": "1.1",
+  "packageFilename": "basic_kbda2.kmp",
+  "jsFilename": "basic_kbda2.js",
+  "encodings": [
+    "unicode"
+  ],
+  "jsFileSize": 16079,
+  "packageFileSize": 53537,
+  "packageIncludes": [
+    "visualKeyboard",
+    "documentation",
+    "welcome"
+  ],
+  "minKeymanVersion": "10.0",
+  "helpLink": "https://help.keyman.com/keyboard/basic_kbda2",
+  "platformSupport": {
+    "windows": "full",
+    "macos": "full",
+    "linux": "full",
+    "desktopWeb": "full",
+    "mobileWeb": "full",
+    "android": "full",
+    "ios": "full"
+  }
+}
