+++ api/test/bundles/fixtures/bundles.test.data.all.expected.json
@@ -1,0 +1,47 @@
+{
+    "acoli": {
+        "1.0": {
+            "windows": {
+                "10.0.1207.0": {
+                    "url": "/keyboards/acoli/1.0/keymandesktop-10.0.1207.0-acoli-1.0.exe"
+                },
+                "11.0.1361.0": {
+                    "url": "/keyboards/acoli/1.0/keymandesktop-11.0.1361.0-acoli-1.0.exe"
+                },
+                "12.0.52.0": {
+                    "url": "/keyboards/acoli/1.0/keymandesktop-12.0.52.0-acoli-1.0.exe"
+                }
+            }
+        }
+    },
+    "africadeadkey": {
+        "1.0": {
+            "windows": {
+                "10.0.1207.0": {
+                    "url": "/keyboards/africadeadkey/1.0/keymandesktop-10.0.1207.0-africadeadkey-1.0.exe"
+                },
+                "11.0.1361.0": {
+                    "url": "/keyboards/africadeadkey/1.0/keymandesktop-11.0.1361.0-africadeadkey-1.0.exe"
+                },
+                "12.0.52.0": {
+                    "url": "/keyboards/africadeadkey/1.0/keymandesktop-12.0.52.0-africadeadkey-1.0.exe"
+                }
+            }
+        }
+    },
+    "basic_kbdinen": {
+        "1.0": {
+            "windows": {
+                "10.0.1207.0": {
+                    "url": "/keyboards/basic_kbdinen/1.0/keymandesktop-10.0.1207.0-basic_kbdinen-1.0.exe"
+                },
+                "11.0.1361.0": {
+                    "url": "/keyboards/basic_kbdinen/1.0/keymandesktop-11.0.1361.0-basic_kbdinen-1.0.exe"
+                },
+                "12.0.52.0": {
+                    "url": "/keyboards/basic_kbdinen/1.0/keymandesktop-12.0.52.0-basic_kbdinen-1.0.exe"
+                }
+            }
+        }
+    }
+}
