+++ api/test/bundles/fixtures/bundles.test.data.current.expected.json
@@ -1,0 +1,74 @@
+{
+    "acoli": {
+        "1.0": {
+            "windows": [
+                {
+                    "url": "/keyboards/acoli/1.0/keymandesktop-12.0.52.0-acoli-1.0.exe"
+                }
+            ]
+        }
+    },
+    "africadeadkey": {
+        "1.0": {
+            "windows": [
+                {
+                    "url": "/keyboards/africadeadkey/1.0/keymandesktop-12.0.52.0-africadeadkey-1.0.exe"
+                }
+            ]
+        }
+    },
+    "africaus": {
+        "1.0": {
+            "windows": [
+                {
+                    "url": "/keyboards/africaus/1.0/keymandesktop-12.0.52.0-africaus-1.0.exe"
+                }
+            ]
+        }
+    },
+    "afro": {
+        "1.2": {
+            "windows": [
+                {
+                    "url": "/keyboards/afro/1.2/keymandesktop-12.0.52.0-afro-1.2.exe"
+                }
+            ]
+        }
+    },
+    "aksarabali_panlex": {
+        "1.1.0": {
+            "windows": [
+                {
+                    "url": "/keyboards/aksarabali_panlex/1.1.0/keymandesktop-12.0.52.0-aksarabali_panlex-1.1.0.exe"
+                }
+            ]
+        }
+    },
+    "amazigh_lat.kmn": {
+        "1.1": {
+            "windows": [
+                {
+                    "url": "/keyboards/amazigh_lat.kmn/1.1/keymandesktop-12.0.52.0-amazigh_lat.kmn-1.1.exe"
+                }
+            ]
+        }
+    },
+    "anii_2015_fr_pack2": {
+        "2015.05.22.2": {
+            "windows": [
+                {
+                    "url": "/keyboards/anii_2015_fr_pack2/2015.05.22.2/keymandesktop-12.0.52.0-anii_2015_fr_pack2-2015.05.22.2.exe"
+                }
+            ]
+        }
+    },
+    "yorubaok": {
+        "1.0": {
+            "windows": [
+                {
+                    "url": "/keyboards/yorubaok/1.0/keymandesktop-12.0.52.0-yorubaok-1.0.exe"
+                }
+            ]
+        }
+    }
+}
