+++ api/test/bundles/fixtures/bundles.test.data.outdated.expected.json
@@ -1,0 +1,11 @@
+{
+    "basic_kbda2": {
+        "1.1": {
+            "windows": {
+                "12.0.52.0": {
+                    "missing": true
+                }
+            }
+        }
+    }
+}
